rename logging.py to logger.py to avoid name conflict
